Added sendQuestion() function
The way the function works is... when you call sendQuestion("Your question", ["answer1", "answer2", "answer etc..."]) it will send a message on messenger and pop up suggested answer buttons to it with the answer1, etc. values. If the user clicks one of the buttons the function will return the value of the button as a string (so for example "answer1") but if the buttons dont show up, like they didnt for Luke bc of having an older version of messenger, or the user just ignores them and types their own message in reply, it will just return that message as a string.